Separate old kds related code to sched_exec file (#5991)
* Draft checkin for kds separation

* Fixed compilation error and dependency

* allignment issue fixed
