+++ src/runtime_src/core/edge/drm/zocl/zocl_ioctl.c
@@ -15,23 +15,9 @@
 
 #include "sched_exec.h"
 #include "zocl_xclbin.h"
-#include "zocl_generic_cu.h"
 #include "zocl_error.h"
 
 extern int kds_mode;
-
-/* TODO: remove this once new KDS is ready */
-int zocl_xclbin_ctx(struct drm_zocl_dev *zdev, struct drm_zocl_ctx *ctx,
-		    struct sched_client_ctx *client);
-
-int zocl_graph_alloc_ctx(struct drm_zocl_dev *zdev, struct drm_zocl_ctx *ctx,
-        struct sched_client_ctx *client);
-int zocl_graph_free_ctx(struct drm_zocl_dev *zdev, struct drm_zocl_ctx *ctx,
-        struct sched_client_ctx *client);
-int zocl_aie_alloc_ctx(struct drm_zocl_dev *zdev, struct drm_zocl_ctx *ctx,
-        struct sched_client_ctx *client);
-int zocl_aie_free_ctx(struct drm_zocl_dev *zdev, struct drm_zocl_ctx *ctx,
-        struct sched_client_ctx *client);
 
 /*
  * read_axlf and ctx should be protected by zdev_xclbin_lock exclusively.
@@ -68,43 +54,19 @@ zocl_read_axlf_ioctl(struct drm_device *ddev, void *data, struct drm_file *filp)
 int
 zocl_ctx_ioctl(struct drm_device *ddev, void *data, struct drm_file *filp)
 {
-	struct drm_zocl_ctx *args = data;
 	struct drm_zocl_dev *zdev = ZOCL_GET_ZDEV(ddev);
 	int ret = 0;
 
 	if (kds_mode == 1) {
-		/* Do not acquire zdev_xclbin_lock like zocl_xclbin_ctx().
+		/* Do not acquire zdev_xclbin_lock like sched_xclbin_ctx().
 		 * New KDS would lock bitstream when open the fist context.
 		 * The lock bitstream would exclude read_axlf_ioctl().
 		 */
 		ret = zocl_context_ioctl(zdev, data, filp);
-		return ret;
 	}
-
-	switch (args->op) {
-
-	case ZOCL_CTX_OP_OPEN_GCU_FD:
-		return zocl_open_gcu(zdev, args, filp->driver_priv);
-
-	case ZOCL_CTX_OP_ALLOC_GRAPH_CTX:
-		return zocl_graph_alloc_ctx(zdev, args, filp->driver_priv);
-
-	case ZOCL_CTX_OP_FREE_GRAPH_CTX:
-		return zocl_graph_free_ctx(zdev, args, filp->driver_priv);
-
-	case ZOCL_CTX_OP_ALLOC_AIE_CTX:
-		return zocl_aie_alloc_ctx(zdev, args, filp->driver_priv);
-
-	case ZOCL_CTX_OP_FREE_AIE_CTX:
-		return zocl_aie_free_ctx(zdev, args, filp->driver_priv);
-
-	default:
-		break;
-	}
-
-	mutex_lock(&zdev->zdev_xclbin_lock);
-	ret = zocl_xclbin_ctx(zdev, args, filp->driver_priv);
-	mutex_unlock(&zdev->zdev_xclbin_lock);
+	else
+		ret = sched_context_ioctl(zdev, data, filp);
+		
 
 	return ret;
 }
